refactors rd_kafka initialization
